Edit Profile Boiler Plate, Toast addition
