+++ client/TrailSeek/src/components/ToastAlert.js
@@ -4,7 +4,7 @@ const ToastAlert = (msg) => {
   return Alert.alert("", msg, [
     {
       text: "Cancel",
-      onPress: () => console.log("Cancel Pressed"),
+      // onPress: () => console.log("Cancel Pressed"),
       style: "cancel",
     },
   ]);
