+++ client/trailseek/src/screens/CreateEventScreen.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { StyleSheet } from "react-native";
 import { useDispatch } from "react-redux";
+import { Toast } from "native-base";
+import { unwrapResult } from "@reduxjs/toolkit";
 
 import { postEvents } from "../app/eventSlice";
 import EventForms from "../components/EventForm";
@@ -9,7 +11,14 @@ const CreateEventScreen = ({ route, navigation }) => {
   const { trailName, trailID } = route.params;
   const dispatch = useDispatch();
   const onSubmitFunc = async (values) => {
-    const response = await dispatch(postEvents({ inputs: values, trailID }));
+    try {
+      const response = await dispatch(postEvents({ inputs: values, trailID }));
+      unwrapResult(response);
+      Toast.show({ text: "Event Created", type: "success", buttonText: "Ok" });
+      navigation.goBack();
+    } catch (e) {
+      ToastAlert(e.message);
+    }
   };
 
   return <EventForms trailName={trailName} onSubmitFunc={onSubmitFunc} />;
