+++ client/trailseek/src/screens/EditEventScreen.js
@@ -1,23 +1,33 @@
+import { unwrapResult } from "@reduxjs/toolkit";
 import React from "react";
 import { StyleSheet } from "react-native";
 import { useDispatch, useSelector } from "react-redux";
+import { Toast } from "native-base";
 
 import { putEvents } from "../app/eventSlice";
 import EventForms from "../components/EventForm";
+import ToastAlert from "../components/ToastAlert";
 
-const EditEventScreen = () => {
+const EditEventScreen = ({ navigation }) => {
   const { eventData, trailName } = useSelector(
     (state) => state.event.currentEvent[0]
   );
   const dispatch = useDispatch();
   const onSubmitFunc = async (values) => {
-    const response = await dispatch(
-      putEvents({
-        inputs: values,
-        trailID: eventData.trailId,
-        eventID: eventData._id,
-      })
-    );
+    try {
+      const response = await dispatch(
+        putEvents({
+          inputs: values,
+          trailID: eventData.trailId,
+          eventID: eventData._id,
+        })
+      );
+      unwrapResult(response);
+      Toast.show({ text: "Event edited", type: "success", buttonText: "Ok" });
+      navigation.goBack();
+    } catch (e) {
+      ToastAlert(e.message);
+    }
   };
 
   return (
